+++ ichub-backend/controllers/fastapi/app.py
@@ -39,10 +39,22 @@ from models.services.twin_management import TwinRead, TwinAspectRead, TwinAspect
 from tools.submodel_type_util import InvalidSemanticIdError
 from tools import InvalidUUIDError
 
+from .routers import (
+    part_management,
+    partner_management,
+    twin_management,
+    submodel_dispatcher,
+    sharing_handler
+)
+
 tags_metadata = [
     {
         "name": "Part Management",
         "description": "Management of part metadata - including catalog parts, serialized parts, JIS parts and batches"
+    },
+    {
+        "name": "Sharing Functionality",
+        "description": "Sharing functionality for catalog part twins - including sharing of parts with business partners and automatic generation of digital twins and submodels"
     },
     {
         "name": "Partner Management",
@@ -60,87 +72,12 @@ tags_metadata = [
 
 app = FastAPI(title="Industry Core Hub Backend API", version="0.0.1", openapi_tags=tags_metadata)
 
-part_management_service = PartManagementService()
-partner_management_service = PartnerManagementService()
-twin_management_service = TwinManagementService()
-part_sharing_shortcut_service = PartSharingShortcutService()
-submodel_dispatcher_service = SubmodelDispatcherService()
-
-@app.get("/part-management/catalog-part/{manufacturer_id}/{manufacturer_part_id}", response_model=CatalogPartRead, tags=["Part Management"])
-async def part_management_get_catalog_part(manufacturer_id: str, manufacturer_part_id: str) -> Optional[CatalogPartRead]:
-    return part_management_service.get_catalog_part(manufacturer_id, manufacturer_part_id)
-
-@app.get("/part-management/catalog-part", response_model=List[CatalogPartReadWithStatus], tags=["Part Management"])
-async def part_management_get_catalog_parts() -> List[CatalogPartReadWithStatus]:
-    return part_management_service.get_catalog_parts()
-
-@app.post("/part-management/catalog-part", response_model=CatalogPartRead, tags=["Part Management"])
-async def part_management_create_catalog_part(catalog_part_create: CatalogPartCreate) -> CatalogPartReadWithStatus:
-    return part_management_service.create_catalog_part(catalog_part_create)
-
-@app.get("/partner-management/business-partner", response_model=List[BusinessPartnerRead], tags=["Partner Management"])
-async def partner_management_get_business_partners() -> List[BusinessPartnerRead]:
-    return partner_management_service.list_business_partners()
-
-@app.get("/partner-management/business-partner/{business_partner_number}", response_model=Optional[BusinessPartnerRead], tags=["Partner Management"])
-async def partner_management_get_business_partner(business_partner_number: str) -> Optional[BusinessPartnerRead]:
-    return partner_management_service.get_business_partner(business_partner_number)
-
-@app.post("/partner-management/business-partner", response_model=BusinessPartnerRead, tags=["Partner Management"])
-async def partner_management_create_business_partner(business_partner_create: BusinessPartnerCreate) -> BusinessPartnerRead:
-    return partner_management_service.create_business_partner(business_partner_create)
-
-@app.get("/partner-management/business-partner/{business_partner_number}/data-exchange-agreement", response_model=List[DataExchangeAgreementRead], tags=["Partner Management"])
-async def partner_management_get_data_exchange_agreements(business_partner_number: str) -> List[DataExchangeAgreementRead]:
-    return partner_management_service.get_data_exchange_agreements(business_partner_number)
-
-@app.get("/twin-management/catalog-part-twin", response_model=List[CatalogPartTwinRead], tags=["Twin Management"])
-async def twin_management_get_catalog_part_twins(include_data_exchange_agreements: bool = False) -> List[CatalogPartTwinRead]:
-    return twin_management_service.get_catalog_part_twins(include_data_exchange_agreements=include_data_exchange_agreements)
-
-@app.get("/twin-management/catalog-part-twin/{global_id}", response_model=List[CatalogPartTwinDetailsRead], tags=["Twin Management"])
-async def twin_management_get_catalog_part_twin(global_id: UUID) -> List[CatalogPartTwinDetailsRead]:
-    return twin_management_service.get_catalog_part_twin_details(global_id)
-
-@app.post("/twin-management/catalog-part-twin", response_model=TwinRead, tags=["Twin Management"])
-async def twin_management_create_catalog_part_twin(catalog_part_twin_create: CatalogPartTwinCreate) -> TwinRead:
-    return twin_management_service.create_catalog_part_twin(catalog_part_twin_create)
-
-@app.post("/twin-management/catalog-part-twin/share", responses={
-    201: {"description": "Catalog part twin shared successfully"},
-    204: {"description": "Catalog part twin already shared"}
-}, tags=["Twin Management"])
-async def twin_management_share_catalog_part_twin(catalog_part_twin_share: CatalogPartTwinShare):
-    if twin_management_service.create_catalog_part_twin_share(catalog_part_twin_share):
-        return JSONResponse(status_code=201, content={"description":"Catalog part twin shared successfully"})
-    else:
-        return JSONResponse(status_code=204, content={"description":"Catalog part twin already shared"})
-
-@app.post("/twin-management/twin-aspect", response_model=TwinAspectRead, tags=["Twin Management"])
-async def twin_management_create_twin_aspect(twin_aspect_create: TwinAspectCreate) -> TwinAspectRead:
-    return twin_management_service.create_twin_aspect(twin_aspect_create)
-
-@app.post("/share/catalog-part", response_model=CatalogPartTwinDetailsRead, tags=["Twin Management"])
-async def twin_management_create_part_sharing_shortcut(catalog_part_twin_share: CatalogPartTwinShare,
-    auto_generate_part_type_information_submodel:bool = True) -> CatalogPartTwinDetailsRead:
-    return part_sharing_shortcut_service.create_catalog_part_sharing_shortcut(
-        catalog_part_twin_share,
-        auto_generate_part_type_information=auto_generate_part_type_information_submodel
-    )
-
-@app.get("/submodel-dispatcher/{semantic_id}/{global_id}/submodel/$value", response_model=Dict[str, Any], tags=["Submodel Dispatcher"])
-@app.get("/submodel-dispatcher/{semantic_id}/{global_id}/submodel", response_model=Dict[str, Any], tags=["Submodel Dispatcher"])
-@app.get("/submodel-dispatcher/{semantic_id}/{global_id}", response_model=Dict[str, Any], tags=["Submodel Dispatcher"])
-async def submodel_dispatcher_get_submodel_content_submodel_value(
-    semantic_id: str,
-    global_id: UUID,
-    edc_bpn: Optional[str] = Header(default=None, alias="Edc-Bpn", description="The BPN of the consumer delivered by the EDC Data Plane"),
-    edc_contract_agreement_id: Optional[str] = Header(default=None, alias="Edc-Contract-Agreement-Id", description="The contract agreement id of the consumer delivered by the EDC Data Plane")
-) -> Dict[str, Any]:
-    try:
-        return submodel_dispatcher_service.get_submodel_content(edc_bpn, edc_contract_agreement_id, semantic_id, global_id)
-    except FileNotFoundError:
-        return JSONResponse(status_code=404, content={"detail": "Submodel not found"})
+## Include here all the routers for the application.
+app.include_router(part_management.router)
+app.include_router(partner_management.router)
+app.include_router(twin_management.router)
+app.include_router(submodel_dispatcher.router)
+app.include_router(sharing_handler.router)
 
 @app.exception_handler(SubmodelNotSharedWithBusinessPartnerError)
 async def submodel_not_shared_with_business_partner_exception_handler(
@@ -173,20 +110,3 @@ async def invalid_uuid_error_exception_handler(
     return JSONResponse(status_code=422, content={"detail": str(exc)})
 
 
-@app.post("/submodel-dispatcher/{semantic_id}/{global_id}/submodel", status_code=204, tags=["Submodel Dispatcher"])
-async def submodel_dispatcher_upload_submodel(
-    semantic_id: str,
-    global_id: UUID,
-    submodel_payload: Dict[str, Any] = Body(..., description="The submodel JSON payload")
-) -> None:
-    return submodel_dispatcher_service.upload_submodel( global_id, semantic_id, submodel_payload)
-
-@app.delete("/submodel-dispatcher/{semantic_id}/{global_id}/submodel", status_code=204, tags=["Submodel Dispatcher"])
-async def submodel_dispatcher_delete_submodel(
-    semantic_id: str,
-    global_id: UUID
-) -> None:
-    try:
-        submodel_dispatcher_service.delete_submodel(global_id, semantic_id)
-    except FileNotFoundError:
-        return JSONResponse(status_code=404, content={"detail": "Submodel not found"})
