+++ ichub-backend/controllers/fastapi/routers/sharing_handler.py
@@ -1,0 +1,40 @@
+#################################################################################
+# Eclipse Tractus-X - Industry Core Hub Backend
+#
+# Copyright (c) 2025 Contributors to the Eclipse Foundation
+#
+# See the NOTICE file(s) distributed with this work for additional
+# information regarding copyright ownership.
+#
+# This program and the accompanying materials are made available under the
+# terms of the Apache License, Version 2.0 which is available at
+# https://www.apache.org/licenses/LICENSE-2.0.
+#
+# Unless required by routerlicable law or agreed to in writing, software
+# distributed under the License is distributed on an "AS IS" BASIS
+# WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND,
+# either express or implied. See the
+# License for the specific language govern in permissions and limitations
+# under the License.
+#
+# SPDX-License-Identifier: Apache-2.0
+#################################################################################
+
+from fastapi import APIRouter, Body, Header
+
+from services.part_sharing_shortcut_service import PartSharingShortcutService
+from models.services.twin_management import (
+    CatalogPartTwinDetailsRead,
+    CatalogPartTwinShare
+)
+
+router = APIRouter(prefix="/share", tags=["Sharing Functionality"])
+part_sharing_shortcut_service = PartSharingShortcutService()
+
+@router.post("/catalog-part", response_model=CatalogPartTwinDetailsRead)
+async def twin_management_create_part_sharing_shortcut(catalog_part_twin_share: CatalogPartTwinShare,
+    auto_generate_part_type_information_submodel:bool = True) -> CatalogPartTwinDetailsRead:
+    return part_sharing_shortcut_service.create_catalog_part_sharing_shortcut(
+        catalog_part_twin_share,
+        auto_generate_part_type_information=auto_generate_part_type_information_submodel
+    )
